+++ src/main/java/frc/robot/commands/AutonLoader.java
@@ -1,6 +1,5 @@
 package frc.robot.commands;
 
-import java.time.Period;
 import java.util.*;
 
 import com.pathplanner.lib.PathPlanner;
